Add precondition for logging
